+++ docs/rst/conf.py
@@ -47,7 +47,7 @@ project = 'signposting link parser'
 year = '2022'
 author = 'Stian Soiland-Reyes'
 copyright = '{0}, {1}'.format(year, author)
-version = release = '0.3.3'
+version = release = '0.4.0'
 
 pygments_style = 'trac'
 templates_path = ['.']